cleanup

+++ code/datums/components/explodable.dm
@@ -16,6 +16,8 @@
 	var/delete_after
 	/// For items, lets us determine where things should be hit.
 	var/equipped_slot
+	/// Whether this component is currently in the process of exploding.
+	var/tmp/exploding = FALSE
 
 /datum/component/explodable/Initialize(devastation_range_override, heavy_impact_range_override, light_impact_range_override, flame_range_override, flash_range_override, uncapped = FALSE, delete_after = EXPLODABLE_DELETE_PARENT)
 	if(!isatom(parent))
@@ -23,6 +25,8 @@
 
 	RegisterSignal(parent, COMSIG_ATOM_ATTACKBY, PROC_REF(explodable_attack))
 	RegisterSignal(parent, COMSIG_ATOM_EX_ACT, PROC_REF(detonate))
+	RegisterSignal(parent, COMSIG_ATOM_TOOL_ACT(TOOL_WELDER), PROC_REF(welder_react))
+	RegisterSignal(parent, COMSIG_ATOM_BULLET_ACT, PROC_REF(projectile_react))
 	if(ismovable(parent))
 		RegisterSignal(parent, COMSIG_MOVABLE_IMPACT, PROC_REF(explodable_impact))
 		RegisterSignal(parent, COMSIG_MOVABLE_BUMP, PROC_REF(explodable_bump))
@@ -40,12 +44,12 @@
 		src.devastation_range = devastation_range
 	if (heavy_impact_range)
 		src.heavy_impact_range = heavy_impact_range
-	if(light_impact_range_override)
-		light_impact_range = light_impact_range_override
-	if(flame_range_override)
-		flame_range = flame_range_override
-	if(flash_range_override)
-		flash_range = flash_range_override
+	if (light_impact_range)
+		src.light_impact_range = light_impact_range
+	if (flame_range)
+		src.flame_range = flame_range
+	if (flash_range)
+		src.flash_range = flash_range
 	src.uncapped = uncapped
 	src.delete_after = delete_after
 
@@ -53,7 +57,6 @@
 	SIGNAL_HANDLER
 	if(!(I.item_flags & IN_STORAGE))
 		return
-
 	check_if_detonate(I)
 
 /datum/component/explodable/proc/explodable_impact(datum/source, atom/hit_atom, datum/thrownthing/throwingdatum)
@@ -71,6 +74,20 @@
 	SIGNAL_HANDLER
 
 	check_if_detonate(target)
+
+/// Welder check. Here because tool_act is higher priority than attackby.
+/datum/component/explodable/proc/welder_react(datum/source, mob/user, obj/item/tool)
+	SIGNAL_HANDLER
+
+	if(check_if_detonate(tool))
+		return
+
+/// Shot by something
+/datum/component/explodable/proc/projectile_react(datum/source, obj/projectile/shot)
+	SIGNAL_HANDLER
+
+	if(shot.damage_type == BURN && shot.damage > 0)
+		detonate()
 
 ///Called when you attack a specific body part of the thing this is equipped on. Useful for exploding pants.
 /datum/component/explodable/proc/explodable_attack_zone(datum/source, damage, damagetype, def_zone, ...)
@@ -100,31 +117,17 @@
 /// Checks if we're hitting the zone this component is covering
 /datum/component/explodable/proc/is_hitting_zone(def_zone)
 	var/obj/item/item = parent
-	var/mob/living/L = item.loc //Get whoever is equipping the item currently
+	var/mob/living/carbon/wearer = item.loc //Get whoever is equipping the item currently
+	if(!istype(wearer))
+		return FALSE
 
-	if(!istype(L))
-		return
+	var/list/worn_items = wearer.get_equipped_items()
 
-	var/obj/item/bodypart/bodypart = L.get_bodypart(check_zone(def_zone))
+	if(!(item in worn_items))
+		return FALSE
 
-	var/list/equipment_items = list()
-	if(iscarbon(L))
-		var/mob/living/carbon/C = L
-		equipment_items += list(C.head, C.wear_mask, C.back, C.gloves, C.shoes, C.glasses, C.ears)
-		if(ishuman(C))
-			var/mob/living/carbon/human/H = C
-			equipment_items += list(H.wear_suit, H.w_uniform, H.belt, H.s_store, H.wear_id)
-		if(ismonkey(C))
-			var/mob/living/carbon/monkey/H = C
-			equipment_items += list(H.w_uniform)
-
-	for(var/bp in equipment_items)
-		if(!bp)
-			continue
-
-		var/obj/item/I = bp
-		if(I.body_parts_covered & bodypart.body_part)
-			return TRUE
+	if(item.body_parts_covered & def_zone)
+		return TRUE
 	return FALSE
 
 
@@ -132,20 +135,36 @@
 	if(!isitem(target))
 		return
 	var/obj/item/I = target
-	if(!I.get_temperature())
-		return
-	detonate() //If we're touching a hot item we go boom
+	if(I.get_temperature() > FIRE_MINIMUM_TEMPERATURE_TO_EXIST)
+		detonate() //If we're touching a hot item we go boom
+		return TRUE
 
 
 /// Expldoe and remove the object
 /datum/component/explodable/proc/detonate()
 	SIGNAL_HANDLER
+	if (exploding)
+		return // If we don't do this and this doesn't delete it can lock the MC into only processing Input, Timers, and Explosions.
 
 	var/atom/bomb = parent
-	explosion(origin = bomb, devastation_range = devastation_range, heavy_impact_range = heavy_impact_range, light_impact_range = light_impact_range, flame_range = flame_range, flash_range = flash_range, ignorecap = uncapped) //epic explosion time
+	var/log = TRUE
+	if(light_impact_range < 1)
+		log = FALSE
+
+	exploding = TRUE
+	explosion(bomb, devastation_range, heavy_impact_range, light_impact_range, flame_range, flash_range, log, ignorecap = uncapped) //epic explosion time
 
 	switch(delete_after)
 		if(EXPLODABLE_DELETE_SELF)
 			qdel(src)
 		if(EXPLODABLE_DELETE_PARENT)
 			qdel(bomb)
+		else
+			addtimer(CALLBACK(src, PROC_REF(reset_exploding)), 0.1 SECONDS)
+
+/**
+ * Resets the expoding flag
+ */
+/datum/component/explodable/proc/reset_exploding()
+	SIGNAL_HANDLER
+	src.exploding = FALSE
